+++ src/matgl/layers/_core.py
@@ -59,6 +59,7 @@ class MLP(nn.Module):
         return f"MLP({', '.join(dims)})"
 
     @property
+    @torch.jit.unused
     def last_linear(self) -> Linear:
         """Return the last linear layer in the network."""
         for layer in reversed(self.layers):
@@ -68,11 +69,13 @@ class MLP(nn.Module):
         raise RuntimeError(msg)
 
     @property
+    @torch.jit.unused
     def depth(self) -> int:
         """Returns depth of MLP."""
         return self._depth
 
     @property
+    @torch.jit.unused
     def in_features(self) -> int:
         """Return input features of MLP."""
         first_layer = self.layers[0]
@@ -80,6 +83,7 @@ class MLP(nn.Module):
         return first_layer.in_features
 
     @property
+    @torch.jit.unused
     def out_features(self) -> int:
         """Returns output features of MLP."""
         for layer in reversed(self.layers):